Actions added to MenuNodes
pointers to functions now added to MenuNodes dubbed Actions.
they can't point to a classes function so they have to be put into a global function and then the MenuNode points to that global function.

Should be all ready to be put into classes as soon as it has been tested with hardware.

Also to note:
I think we'll be okay with the scope of our project but it's interesting to know that a pointer apparently uses 2 bytes of memory. I doubt we'd use enough pointers for that to be a problem though.